vagrant: use ceph-node as ceph-admin
We always bring-up ceph-admin node with `vagrant up` everytime we
use this cluster, but once the cluster installation is done the
admin node will be idle for ever.

So, just to save resources, merging the ceph-deploy related stuff
into one of the ceph-nodes.

The comments in the Vagrant file should guide to spin a dedicated
ceph-admin node anyday.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>